techdebt: [PIE-4710]: code refactoring for independent deployment - debezium service (#36195)
* code refactoring

* refactoring

* small fix

* changed path in portal-openjdk-bazel.sh

* renamed debezium service

* edited scripts

* added scripts

* fixed code format

* added scripts

* added build.properties file

* fix build

* fix build

* fix build

* fix buidl

* fix build

* fix build

* fix build

* fix build

* fix deployment

* added release folder in debezium service

* removed platform jenkins
