+++ debezium-service/build.properties
@@ -1,0 +1,4 @@
+build.majorVersion=1
+build.minorVersion=0
+build.number=100
+build.patch=000
