+++ debezium-service/container/scripts/replace_configs.sh
