+++ debezium-service/container/scripts/run.sh
