+++ debezium-service/container/scripts/set_default_variables.sh
