+++ debezium-service/container/scripts/start_process.sh
