+++ debezium-service/contracts/BUILD.bazel
@@ -1,0 +1,15 @@
+load("//:tools/bazel/harness.bzl", "java_library")
+load("//:tools/bazel/GenTestRules.bzl", "run_tests")
+load("//:tools/bazel/macros.bzl", "run_analysis")
+
+java_library(
+    name = "module",
+    srcs = glob(["src/main/java/**/*.java"]),
+    resources = [],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//980-commons:module",
+    ],
+)
+
+run_analysis()
