+++ debezium-service/contracts/src/main/java/io/harness/health/HealthResource.java
@@ -7,19 +7,12 @@
 
 package io.harness.health;
 
-import static io.harness.exception.WingsException.USER;
-import static io.harness.maintenance.MaintenanceController.getMaintenanceFlag;
-
 import io.harness.annotations.ExposeInternalException;
-import io.harness.eraro.ErrorCode;
-import io.harness.exception.NoResultFoundException;
 import io.harness.rest.RestResponse;
 import io.harness.security.annotations.PublicApi;
 
 import com.codahale.metrics.annotation.ExceptionMetered;
 import com.codahale.metrics.annotation.Timed;
-import com.codahale.metrics.health.HealthCheck;
-import com.google.inject.Inject;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import javax.ws.rs.Consumes;
@@ -27,42 +20,17 @@ import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
-import lombok.extern.slf4j.Slf4j;
 
 @Api("health")
 @Path("/health")
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
 @PublicApi
-@Slf4j
 @ExposeInternalException
-public class HealthResource {
-  private HealthService healthService;
-
-  @Inject
-  public HealthResource(HealthService healthService) {
-    this.healthService = healthService;
-  }
-
+public interface HealthResource {
   @GET
   @Timed
   @ExceptionMetered
   @ApiOperation(value = "get health for Debezium service", nickname = "getDebeziumHealthStatus")
-  public RestResponse<String> get() throws Exception {
-    if (getMaintenanceFlag()) {
-      log.info("In maintenance mode. Throwing exception to prevent traffic.");
-      throw NoResultFoundException.newBuilder()
-          .code(ErrorCode.RESOURCE_NOT_FOUND)
-          .message("in maintenance mode")
-          .reportTargets(USER)
-          .build();
-    }
-
-    final HealthCheck.Result check = healthService.check();
-    if (check.isHealthy()) {
-      return new RestResponse<>("healthy");
-    }
-
-    throw new HealthException(check.getMessage(), check.getError());
-  }
+  RestResponse<String> get() throws Exception;
 }
