+++ debezium-service/service/BUILD.bazel
@@ -6,12 +6,13 @@ java_binary(
     srcs = glob(["src/main/java/**/*.java"]),
     main_class = "io.harness.DebeziumServiceApplication",
     resources = [
-        "//951-debezium-service/src/main/resources:resource",
+        "//debezium-service/service/src/main/resources:resource",
     ],
-    visibility = ["//visibility:public"],
+    visibility = ["//debezium-service/build:__subpackages__"],
     deps = [
         "//:lombok",
         "//952-debezium-engine:module",
+        "//debezium-service/contracts:module",
         "@maven//:ch_qos_logback_logback_access",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:ch_qos_logback_logback_core",
