+++ debezium-service/service/src/main/resources/BUILD.bazel
