+++ debezium-service/service/src/main/resources/META-INF/services/io.dropwizard.logging.AppenderFactory
