+++ debezium-service/service/src/test/resources/BUILD.bazel
